Cell component adjustments
