+++ src/components/TableComponent/Cell/Cell.tsx
@@ -1,0 +1,35 @@
+import React from 'react';
+import './Cell.scss';
+import Identicon from '@polkadot/react-identicon';
+import { TableData } from '../../../types/types';
+
+
+const Cell: React.FC<{ cell: TableData }> = ({ cell }) => {
+
+  const formatAccountString = (account: string): string => {
+    if (account.length <= 19) return account;
+    const firstPart = account.slice(0, 6);
+    const lastPart = account.slice(-4);
+    return `${firstPart}...${lastPart}`;
+  };
+
+  switch (cell.cellType) {
+    case 'text':
+      return <p>{cell.data}</p>;
+    case 'link':
+      return <a target='_target' href={cell.link}>{cell.data}</a>;
+    case 'address': 
+      return (
+        <>
+          <span className='addressCell-icons'>
+            <Identicon value={cell.data} size={20} className="identicon"/>
+          </span>
+          <a target='_target' href={cell.link}>{formatAccountString(cell.data)}</a>
+        </>
+      );
+    default:
+      return null;
+  }
+};
+
+export default Cell;
