+++ src/styles/global.scss
@@ -17,3 +17,7 @@
   --gray3: #{$gray3};
   --greenPrimary: #{$greenPrimary};
 }
+
+a {
+  color: $greenPrimary;
+}
